pyproject.toml
[tool.poetry]
name = "make-responsive-images"
version = "0.1.9"
description = "Generate responsive images for your website, so you can use srcset in your <img> tags and serve an optimal image to each device that views your site."
readme = "README.md"
authors = ["Sean McCarthy <sean.mccarthy@hotmail.com>"]
license = "MIT"
repository = "https://github.com/mccarthysean/make-responsive-images"
homepage = "https://mccarthysean.dev/make-responsive-images"
documentation = "https://mccarthysean.dev/make-responsive-images"
keywords = ["responsive", "images", "generator"]

[tool.poetry.urls]
"Sean's Blog" = "https://mccarthysean.dev"
"Sean's Company" = "https://myijack.com"

[tool.poetry.scripts]
resize = "make_responsive_images.main:app"

[tool.poetry.dependencies]
python = "^3.6.2"
Pillow = "^8.4.0"
typer = "0.3.2"
colorama = "^0.4.4"
shellingham = "^1.4.0"

[tool.poetry.dev-dependencies]
bandit = "^1.7.0"
safety = "^1.10.3"
dodgy = "^0.2.1"
black = "^21.10b0"
autoflake = "^1.4"
isort = "^5.10.1"
pylint = "^2.11.1"
mypy = "^0.782"
flake8 = "^3.7.9"
pre-commit = "^2.4.0"
pytest = "^6.0.1"
pytest-cov = "^2.8.1"
mkdocs = "^1.1"
mkdocs-material = "^7.3.6"
markdown-include = "^0.6.0"
Werkzeug = "^2.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
